Remove whitespace at end of lines
It was actually there on purpose.

I used to use the w flag in vim's 'formatoptions' option. This flag
means the line was not terminated, for the purposes of the text
wrapping functionality, and so intentional line breaks could be told
apart from automatic ones. But emails and git don't like this, so I no
longer use it.

> 							*fo-table*
...
> letter	 meaning when present in 'formatoptions'    ~
...
> w	Trailing white space indicates a paragraph continues in the next line.
> 	A line that ends in a non-white character ends a paragraph.